+++ addon/button.js
@@ -7,7 +7,8 @@
 // auraLoadingBox: Lightning / Salesforce1
 // studioBody: Exoperience Builder
 // flowContainer: Flow Debugger
-if (document.querySelector("body.sfdcBody, body.ApexCSIPage, #auraLoadingBox, #studioBody, #flowContainer") || location.host.endsWith("visualforce.com")) {
+const visualForceDomains = ["visualforce.com", "vf.force.com"];
+if (document.querySelector("body.sfdcBody, body.ApexCSIPage, #auraLoadingBox, #studioBody, #flowContainer") || visualForceDomains.filter(host => location.host.endsWith(host)).length > 0) {
   // We are in a Salesforce org
   chrome.runtime.sendMessage({message: "getSfHost", url: location.href}, sfHost => {
     if (sfHost) {